app: implement pipes (NOTE, please see full message)
The pipe implementation *relies* on each individual command honouring
the `varin` argument, which it does not yet do. So, pipes will "work",
but the piped data will be ignored because I haven't coded any of the
downstream behaviour. I've only coded the part which allows passing of
varout to varin, the commands still don't do anything with varin.